+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0
+
+- Change @param 'colorCode' to 'colorName'
+- Make @param chunkSize optional preventing unexpected line breaks, giving more flexibility to the user
+
 ## 1.0.2
 
 - Add 'colorizeText' function to colorize text before printing it to console